fix: resolve atomic interface conflict
