+++ go.mod
@@ -11,7 +11,7 @@ require (
 	github.com/zoobzio/atom v1.0.0
 	github.com/zoobzio/edamame v1.0.1
 	github.com/zoobzio/sentinel v1.0.2
-	github.com/zoobzio/soy v1.0.1
+	github.com/zoobzio/soy v1.0.2
 	github.com/zoobzio/vecna v0.0.2
 )
